+++ react_frontend/package.json
@@ -17,11 +17,13 @@
     "react-error-boundary": "3.1.0",
     "react-responsive-modal": "6.0.0",
     "react-use-clipboard": "1.0.2",
+    "rfc4648": "1.4.0",
     "swr": "0.3.9",
     "tailwindcss": "1.9.6"
   },
   "devDependencies": {
     "@testing-library/react": "11.2.2",
+    "babel-eslint": "10.1.0",
     "eslint": "7.14.0",
     "eslint-config-airbnb": "18.2.1",
     "eslint-config-prettier": "6.15.0",
@@ -43,6 +45,7 @@
       "browser": true,
       "jest": true
     },
+    "parser": "babel-eslint",
     "extends": [
       "airbnb",
       "airbnb/hooks",
